requirements.txt
flask>=2.0.0
gunicorn>=20.1.0
flask-cors
langchain
langchain-community
langchain-google-genai
python-dotenv
google-search-results 
joblib
scikit-learn
qdrant_client